requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai==0.8.3
python-multipart==0.0.9
ezdxf==1.3.5  # For DXF exports (optional for MVP)
jinja2==3.1.4  # For templating if needed
Pillow==10.4.0  # For PIL Image support
httpx==0.27.0  # For HTTP client (IP detection)